move vitals logic to VitalsPageStore
